refactor(#169): modify hot posts -> mentor-board, remove controller
